home-encrypted-backup-doc-system

+++ .Edit_Stop.command
@@ -4,18 +4,22 @@ HEBDS_BACKUP_PATH=~/Google\ Drive/encrypt_backup
 cd $HEBDS_DIR_PATH
 
 tar -zcvf encrypt_dir.tar.gz ./encrypt_dir
-rm -rf ./encrypt_dir
 openssl aes-256-cbc -in encrypt_dir.tar.gz -out encrypt_dir.aes
 rm -r encrypt_dir.tar.gz
 
-git add .
-epochTime=$(date '+%Y%m%d %H%M%S')
-git commit -m "${epochTime}"
+if [ -e encrypt_dir.aes ]
+then
+    rm -rf ./encrypt_dir
 
-cp -rf encrypt_dir.aes "${HEBDS_BACKUP_PATH}/encrypt_dir.aes"
-cp -rf .git "${HEBDS_BACKUP_PATH}/.git"
+    git add .
+    epochTime=$(date '+%Y%m%d %H%M%S')
+    git commit -m "${epochTime}"
 
-mv Edit_Stop.command .Edit_Stop.command
-mv .Edit_Start.command Edit_Start.command
-mv .Read_Start.command Read_Start.command
-mv .Pull_Sync.command Pull_Sync.command
+    cp -rf encrypt_dir.aes "${HEBDS_BACKUP_PATH}/encrypt_dir.aes"
+    cp -rf .git "${HEBDS_BACKUP_PATH}/.git"
+
+    mv Edit_Stop.command .Edit_Stop.command
+    mv .Edit_Start.command Edit_Start.command
+    mv .Read_Start.command Read_Start.command
+    mv .Pull_Sync.command Pull_Sync.command
+fi
